+++ analysis/import/SVNLogParser/src/test/kotlin/de/maibornwolff/codecharta/importer/svnlogparser/input/metrics/AbsoluteCodeChurnTest.kt
@@ -2,14 +2,14 @@ package de.maibornwolff.codecharta.importer.svnlogparser.input.metrics
 
 import de.maibornwolff.codecharta.importer.svnlogparser.input.Modification
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
 class AbsoluteCodeChurnTest {
 
     private val FILENAME = "filename"
 
     @Test
-    fun should_have_initial_value_zero() {
+    fun `should have initial value zero`() {
         // when
         val metric = AbsoluteCodeChurn()
 
@@ -18,7 +18,7 @@ class AbsoluteCodeChurnTest {
     }
 
     @Test
-    fun should_increase_by_single_modification() {
+    fun `should increase by single modification`() {
         // given
         val metric = AbsoluteCodeChurn()