+++ analysis/model/src/main/kotlin/de/maibornwolff/codecharta/model/NodeMaxAttributeMerger.kt
@@ -59,6 +59,6 @@ class NodeMaxAttributeMerger(var mergeChildrenList: Boolean = false) : NodeMerge
         }
     }
 
-    private fun <K, V> Map<K, V>.mergeReduce(other: Map<K, V>, reduce: (V, V) -> V = { _, b -> b }): Map<K, V> =
+    private fun <K, V : Any> Map<K, V>.mergeReduce(other: Map<K, V>, reduce: (V, V) -> V = { _, b -> b }): Map<K, V> =
         this.toMutableMap().apply { other.forEach { merge(it.key, it.value, reduce) } }
 }